server/pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "odin-tiktok-slide-server"
version = "0.1.0"
description = "FastAPI backend for ODIN TikTok Slide"
authors = [{ name = "ODIN Team" }]
dependencies = [
    "fastapi>=0.115.0,<0.116.0",
    "uvicorn[standard]>=0.30.0,<0.31.0",
    "pydantic>=2.7,<3.0",
    "python-dotenv>=1.0,<2.0",
    "google-genai>=0.3,<1.0",
    "pydantic-settings>=2.0,<3.0",
    "python-multipart>=0.0.9,<1.0",
    "Pillow>=10.0,<11.0",
    "numpy>=1.26,<3.0",
    "opencv-python>=4.8,<5.0",
    "requests>=2.28,<3.0",
    "bcrypt>=4.1,<5.0"
]
requires-python = ">=3.10"

[project.optional-dependencies]
dev = ["ruff>=0.5,<0.6", "pytest>=8.0,<9.0"]

[tool.uvicorn]
factory = false
reload = true
host = "0.0.0.0"
port = 8000
